Improved project by removing console logs, renaming files, and adding a live version link to the README
