+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"C:\\Users\\front\\Desktop\\webdev\\react\\react-redux-bookstore\\src\\actions\\index.js":"1","C:\\Users\\front\\Desktop\\webdev\\react\\react-redux-bookstore\\src\\containers\\BooksList.js":"2"},{"size":326,"mtime":1611777056157,"results":"3","hashOfConfig":"4"},{"size":2088,"mtime":1611777156993,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"qlxl2s",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\front\\Desktop\\webdev\\react\\react-redux-bookstore\\src\\actions\\index.js",[],"C:\\Users\\front\\Desktop\\webdev\\react\\react-redux-bookstore\\src\\containers\\BooksList.js",[]]
